[DF-6727] Fix warnings and disabled mno-thumb for iOS.
Former-commit-id: f3baa816be41e0937bfad7a7a3f6f31423b6b21d

+++ Sources/Internal/Render/2D/TextLayout.cpp
@@ -133,11 +133,11 @@ void TextLayout::NextByWords(const float32 lineWidth)
     float32 targetWidth = std::floor(lineWidth);
     float32 currentWidth = 0;
     uint32 textLength = (uint32)preparedText.length();
-    uint32 lastPossibleBreak = 0;
+    size_t lastPossibleBreak = 0;
 
     DVASSERT_MSG(textLength == breaks.size(), "Reset with wrong wrap mode");
 
-    for (uint32 pos = fromPos; pos < textLength; ++pos)
+    for (size_t pos = fromPos; pos < textLength; ++pos)
     {
         char16 ch = preparedText[pos];
         uint8 canBreak = breaks[pos];
@@ -188,10 +188,10 @@ void TextLayout::NextBySymbols(const float32 lineWidth)
 {
     float32 targetWidth = std::floor(lineWidth);
     float32 currentLineDx = 0;
-    int32 totalSize = preparedText.length();
-    int32 currentLineEnd = 0;
+    size_t totalSize = preparedText.length();
+    size_t currentLineEnd = 0;
 
-    for (int pos = fromPos; pos < totalSize; pos++)
+    for (size_t pos = fromPos; pos < totalSize; pos++)
     {
         char16 t = preparedText[pos];
         char16 tNext = 0;
